requirements.txt
fastapi
uvicorn[standard]
transformers
torch  # pick the correct torch wheel for your system (CUDA vs CPU)
Pillow
requests
scikit-image
pydantic
accelerate
bitsandbytes
python-multipart
qwen-vl-utils